refactor: move generator validation to core-blockchain
